+++ packages/atproto/lib/src/atproto.dart
@@ -7,7 +7,7 @@ import 'package:atproto_core/atproto_core.dart' as core;
 
 import 'atproto_service.dart';
 import 'entities/session.dart';
-import 'handles/handles_service.dart';
+import 'identities/identities_service.dart';
 import 'repositories/repositories_service.dart';
 import 'sessions/sessions_service.dart';
 
@@ -50,8 +50,8 @@ abstract class ATProto {
   /// Returns the sessions service.
   SessionsService get sessions;
 
-  /// Returns the handles service.
-  HandlesService get handles;
+  /// Returns the identities service.
+  IdentitiesService get identities;
 
   /// Returns the repositories service.
   RepositoriesService get repositories;
@@ -85,7 +85,7 @@ class _ATProto implements ATProto {
   SessionsService get sessions => _service.sessions;
 
   @override
-  HandlesService get handles => _service.handles;
+  IdentitiesService get identities => _service.identities;
 
   @override
   RepositoriesService get repositories => _service.repositories;